+++ state/atoms/selectedPatient.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const selectedPatientState = atom<string | null>({
+  key: 'selectedPatientState',
+  default: null
+});
